+++ cmd/brabble/main.go
@@ -36,6 +36,7 @@ func run() error {
 	root.AddCommand(control.NewDoctorCmd(cfgPath))
 	root.AddCommand(control.NewServiceCmd(cfgPath))
 	root.AddCommand(control.NewSetupCmd(cfgPath))
+	root.AddCommand(control.NewReloadCmd(cfgPath))
 
 	// Hidden internal serve command used by start.
 	root.AddCommand(daemon.NewServeCmd(cfgPath))